refactor: change module and function names to be more indicative
Refs: BLAIS5-3048
