+++ src/components/instrumentDetails/sections/viewCatiModeDetails.test.tsx
@@ -6,7 +6,7 @@ import flushPromises from "../../../tests/utils";
 import { render, waitFor, screen } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import React from "react";
-import ViewToStartDate from "./viewToStartDate";
+import ViewCatiModeDetails from "./viewCatiModeDetails";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 
@@ -25,7 +25,7 @@ describe("View TO Start Date section", () => {
 
     it("should display an error message when it fails to load the TO Start Date", async () => {
         render(
-            <ViewToStartDate instrumentName={"OPN2101A"} />
+            <ViewCatiModeDetails instrumentName={"OPN2101A"} />
         );
 
         await act(async () => {
